==> 2_GrandRiverMB_alphadiversity_MEsser.Rmd <==
---
title: "Grand River 2022 - microbiome alpha diversity analysis"
author: "Milena Esser"
date: "2025-05-20"
output:
  html_document: 
    toc: true
    toc_depth: 4
    number_sections: true
    toc_float: true
  pdf_document: default
editor_options:
  chunk_output_type: console
---

## Set up

```{r 2-setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
rm(list=ls()) #Clears environment
#Change margin size to be smaller so graphs fit in the plot panel
par(mar = c(2, 2, 2, 2)) # Set the margin on all sides to 2

###Loading packages
library(phyloseq) #phyloseq will be the main package used for structuring microbiome data and diversity comparisons
library(ggplot2) #For creating graphs
library(ggtext)
library(plyr) #if you load this after dpylr it messes up the dplyr functions
library(dplyr) #Helps with data wrangling 
library(vegan)
library(tidyverse) #data wrangling
library(knitr) #For R Markdown knitting
library(RColorBrewer)
library(ggpubr)
library(MicEco) #to prune based on prevalence
library(picante)
library(ggh4x) # For facet_nested_wrap()
library(rstatix)
library(tidyr)
library(purrr)
library(readr)

# Create output directory if it doesn't exist
output_dir <- "R_output/Alpha_diversity"
if (!dir.exists(output_dir)) {
  dir.create(output_dir, recursive = TRUE)
}
```

## Loading Data

Loading the phyloseq objects

```{r Loading-data}
#not rarefied dataset "ps_final"
load("R_output//ps_objects//GrandRiverMB_phyloseq_filt_unnorm.RData")

#rarefied dataset "ps_rare"
load("R_output//ps_objects//GrandRiverMB_phyloseq_filt_rar.RData")

```

## Data wrangling & calc

```{r data-wrangling-and-calc, echo=TRUE}
#PDtree <- phy_tree(ps_rare)
otu <- otu_table(ps_final)
otu <- as.data.frame(t(otu))

alpha.pd.sr<-otu


#Adding vector to summarize replicates
alpha.pd.sr$site <- sample_data(ps_final)$Site_ID
alpha.pd.sr$taxa <- sample_data(ps_final)$Taxa
alpha.pd.sr$lifestage <- sample_data(ps_final)$Lifestage
alpha.pd.sr$location<- sample_data(ps_final)$Location
alpha.pd.sr$city<- sample_data(ps_final)$Site
alpha.pd.sr$sample_type<- sample_data(ps_final)$Sample_Type
#---------------------------------------------------
#### Eveness
alpha_diversity <- estimate_richness(ps_final, measure = c("Shannon", "Observed"))
alpha_diversity
H <- alpha_diversity$Shannon
S1 <- alpha_diversity$Observed
S <- log(S1)
eveness <- H/S
#eveness
alpha.pd.sr$Evenness = eveness


alpha.pd.sr$site<-factor(alpha.pd.sr$site, levels=c("WAT_DWN","WAT_UP","HESP_UP","HESP_DWN","CAL_UP","CAL_DWN","KIT_UP", "KIT_DWN" ))

alpha.pd.sr$taxa<-factor(alpha.pd.sr$taxa, levels=c("Hydropsychidae", "Chironomidae", "Heptageniidae", "Araneid", "Tetragnatha"))

alpha.pd.sr$location<-factor(alpha.pd.sr$location, levels=c("UP", "DWN" ))
```

## Plots

```{r Plot-all}


# Plot
p_all <- ggplot(data = alpha.pd.sr, aes(x = location, y = Evenness, fill = site)) + 
  facet_nested_wrap(~ city + lifestage + taxa, scales = "free_x", nrow = 1) +
  geom_boxplot(position = position_dodge2(preserve = "single")) +
  
  # Labels
  labs(x = "Sample", y = "Evenness") +
  
  # Theme
  theme_bw(base_size = 13) +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1, face = "bold", size = 10),
    axis.title.x = element_text(margin = margin(t = 6), face = "bold", size = 14),
    axis.title.y = element_text(margin = margin(r = 6), face = "bold", size = 14),
    axis.text.y = element_text(size = 12),
    axis.text = element_text(face = "italic"),
    strip.text.x = element_text(size = 13, face = "bold"),
    legend.position = "right"
  ) +
  
  # Color palette
  scale_fill_brewer(palette = "Accent")

# Show plot
p_all
#save plot
ggsave(filename = file.path(output_dir, "Evenness_all.pdf"), height = 6, width = 24)
ggsave(filename = file.path(output_dir, "Evenness_all.png"), height = 6, width = 24)

```

```{r Plot-city-by-city}
#plot each city individually
city_list <- unique(alpha.pd.sr$city)

for (cty in city_list) {
  p <- ggplot(subset(alpha.pd.sr, city == cty), 
              aes(x = location, y = Evenness, fill = site)) +
    facet_nested(~ lifestage + taxa, scales = "free_x") +
    geom_boxplot(position = position_dodge2(preserve = "single")) +
    labs(#title = paste("City:", cty),
         x = "Location",
         y = "Evenness",
         fill = "Site") +
    theme_bw(base_size = 13) +
    theme(
      axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1, face = "bold", size = 9),
      strip.text.x = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "italic"),
      plot.title = element_text(face = "bold", size = 14, hjust = 0.5),
      legend.position = "none"
    ) +
    scale_fill_brewer(palette = "Accent")
  
  print(p)
  # save each plot:
ggsave(filename = file.path(output_dir, paste0("Evenness_", cty, ".png")), plot = p, width = 12, height = 5)
ggsave(filename = file.path(output_dir, paste0("Evenness_", cty, ".pdf")), plot = p, width = 12, height = 5)
} 
#---------------------------------------------------    
#Shannon/Simson/invsimson in vegan
#--------------------------------------------------- 

# Calculate diversity indices
alpha.pd.sr$Shannon <- diversity(otu, index = "shannon")
alpha.pd.sr$Simpson <- diversity(otu, index = "simpson")
alpha.pd.sr$InvSimpson <- 1 / alpha.pd.sr$Simpson


#---------------------------------------------------    
#plot
#---------------------------------------------------   
# Plot
p_all <- ggplot(data = alpha.pd.sr, aes(x = location, y = Shannon, fill = site)) + 
  facet_nested_wrap(~ city + lifestage + taxa, scales = "free_x", nrow = 1) +
  geom_boxplot(position = position_dodge2(preserve = "single")) +
  
  # Labels
  labs(x = "Sample", y = "Shannon") +
  
  # Theme
  theme_bw(base_size = 13) +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1, face = "bold", size = 10),
    axis.title.x = element_text(margin = margin(t = 6), face = "bold", size = 14),
    axis.title.y = element_text(margin = margin(r = 6), face = "bold", size = 14),
    axis.text.y = element_text(size = 12),
    axis.text = element_text(face = "italic"),
    strip.text.x = element_text(size = 13, face = "bold"),
    legend.position = "right"
  ) +
  
  # Color palette
  scale_fill_brewer(palette = "Accent")

# Show plot
p_all
#save plot
ggsave(filename = file.path(output_dir, "Shannon_all.pdf"), height = 6, width = 24)
ggsave(filename = file.path(output_dir, "Shannon_all.png"), height = 6, width = 24)


#plot each city individually
city_list <- unique(alpha.pd.sr$city)

for (cty in city_list) {
  p <- ggplot(subset(alpha.pd.sr, city == cty), 
              aes(x = location, y = Shannon, fill = site)) +
    facet_nested(~ lifestage + taxa, scales = "free_x") +
    geom_boxplot(position = position_dodge2(preserve = "single")) +
    labs(#title = paste("City:", cty),
         x = "Location",
         y = "Shannon",
         fill = "Site") +
    theme_bw(base_size = 13) +
    theme(
      axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1, face = "bold", size = 9),
      strip.text.x = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "italic"),
      plot.title = element_text(face = "bold", size = 14, hjust = 0.5),
      legend.position = "none"
    ) +
    scale_fill_brewer(palette = "Accent")
  
  print(p)
  # save each plot:
ggsave(filename = file.path(output_dir, paste0("Shannon_", cty, ".png")), plot = p, width = 12, height = 5)
ggsave(filename = file.path(output_dir, paste0("Shannon_", cty, ".pdf")), plot = p, width = 12, height = 5)

}

```

## Statistics - Wilcoxon

checking if diversity differed significantly between sites

```{r Statistical-tests, echo=TRUE, cache=TRUE}
alpha.pd.sr %>% 
  group_by(city, location, sample_type) %>%
  get_summary_stats(Shannon, type = "common")

# Filter out groups with only one site and nest data for grouped testing
alpha_grouped <- alpha.pd.sr %>%
  group_by(city, sample_type) %>%
  filter(n_distinct(site) > 1) %>%
  nest()

# Apply Wilcoxon tests across groups using purrr::map
wilcox_results <- alpha_grouped %>%
  mutate(
    pwc_Even = map(data, ~ wilcox_test(.x, Evenness ~ site, p.adjust.method = "BH")),
    pwc_Shan = map(data, ~ wilcox_test(.x, Shannon ~ site, p.adjust.method = "BH"))
  )

# Unnest and write output
pwc_Even <- wilcox_results %>%
  select(city, sample_type, pwc_Even) %>%
  unnest(pwc_Even)

pwc_Shan <- wilcox_results %>%
  select(city, sample_type, pwc_Shan) %>%
  unnest(pwc_Shan)

# Export results to CSV
readr::write_csv(pwc_Even, file.path(output_dir, "wilcoxon_results_evenness.csv"))
readr::write_csv(pwc_Shan, file.path(output_dir, "wilcoxon_results_shannon.csv"))

# Save full alpha diversity data (optional)
readr::write_csv(alpha.pd.sr, file.path(output_dir, "alphadiversity_results.csv"))

```
